@@ -1,15 +1,16 @@
 import { Observable } from 'rxjs/Observable';
-import { Response } from '@angular/http'
+import { HttpErrorResponse } from '@angular/common/http'
 import 'rxjs/add/observable/throw'
 
 export class ErrorHandler {
-    static handlerError(error: Response | any){
+    static handlerError(error: HttpErrorResponse | any){
         let errorMessage: string
 
-        if(error instanceof Response){
-            errorMessage = `Erro ${error.status} ao obter a URL ${error.url} - ${error.statusText}`
+        if(error instanceof HttpErrorResponse){
+            const body = error.error
+            errorMessage = `Erro ${error.url} ao obter a URL ${error.status} - ${error.statusText || ''} ${body}`
         }else{
-            errorMessage = error.toString()
+            errorMessage = error.message ? error.message : error.toString()
         }
 
         console.log(errorMessage)
