@@ -3,7 +3,7 @@ import { Observable } from 'rxjs/Observable';
 import { ShoppingCartService } from './../restaurant-detail/shopping-cart/shopping-cart.service';
 import { Injectable } from "@angular/core";
 
-import {Http, Headers, RequestOptions} from '@angular/http'
+import {HttpClient} from '@angular/common/http'
 import { Order, OrderItem } from './order.model'
 import 'rxjs/add/operator/map'
 
@@ -12,7 +12,7 @@ import { CartItem } from '../restaurant-detail/shopping-cart/cart-item.model';
 @Injectable()
 export class OrderService {
     //ShoppingCartService que manipula os dados dos itens do carrinho
-    constructor(private cartService: ShoppingCartService, private http: Http){}
+    constructor(private cartService: ShoppingCartService, private http: HttpClient){}
 
     itemsValue(): number{
         return this.cartService.total()
@@ -39,12 +39,7 @@ export class OrderService {
     }
 
     checkOrder(order: Order): Observable<string> { //como envia json que é uma representação textual, preciso converter
-        const headers = new Headers()
-        headers.append('Content-Type', 'application/json')  //formato que vamos mandar no corpo do request
-        return this.http.post(`${MEAT_API}/orders`, 
-        JSON.stringify(order),
-        new RequestOptions({headers: headers}))
-        .map(response => response.json())  //queremos apenas o corpo, não queremos status e tudo mais que temos de reposta, por isso o map
+        return this.http.post<Order>(`${MEAT_API}/orders`, order)
         .map(order => order.id)
     }
 }
